Move InMemInstancesRepository
