+++ collect_app/src/test/java/org/odk/collect/android/activities/FormMapViewModelTest.java
@@ -9,7 +9,7 @@ import org.odk.collect.android.activities.viewmodels.FormMapViewModel;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.instances.Instance;
 import org.odk.collect.forms.instances.InstancesRepository;
-import org.odk.collect.android.support.InMemInstancesRepository;
+import org.odk.collect.formstest.InMemInstancesRepository;
 
 import java.util.Arrays;
 import java.util.List;
